+++ decidim-consultations/app/assets/javascripts/decidim/consultations/social_share.js
@@ -1,0 +1,2 @@
+// = require social-share-button
+// = require_self